greeter: Make wallpapers work in Plasma 6
In Plasma 6 wallpapers instantiate WallpaperItem as root item
themselves. So we need to let our wrapper gain the full API, make
it an item and register it to be instantiated. Because pluginName
and configuration are CONSTANT upstream we need to make sure they
are already correct when the item is complete.

BUG:472876
CCBUG:470601