greeter: disable qml disk cache
The cache currently runs risk of getting corrupted and breaking the lock
screen until the cache hash changes or the cache file gets removed by
the user. Since we don't necessarily need a cache here let's just
disable it for the time being in the interest of greater fault
tolerance.

BUG: 471952