Reset lock complete variable after lock
adfae58490b4b2307221fa4e45465948b749937b changed the greeter to notify
out of band that the lock has been successfully lifted.

s_lockProcessRequestedExit represents the state between unlock occurring
and the greeter exiting. This was not reset when we next showed the
lockscreen.

A symptom amongst others is that the grace locker would be broken if
used after a successful lock.

BUG: 476567