greeter: Don't cancel authentications when a non-interactive auth fails
A non-interactive authentication like fingerprint should not cancel
other authentication methods.

It also does not make sense to mark the entire authentication as failed.
PamAuthenticators abstracts the concept of multiple backends. We are in
a succeeded state if any succeed. We are not in a failed state when one
fails.