KCM: make delay unit not editable and overhaul delay-setting UI
Currently the user can select and edit the delay units in the config,
leading them to believe that they can change "minutes" to "seconds"
or "hours". This is not the case, and causes confusion.

See this Reddit thread for reference of the confusion:
https://old.reddit.com/r/kde/comments/tpvp5h/why_such_an_arbitrary_number_for_allow_unlocking/

This change makes them ComboBox components instead, with an optional
"Custom" entry for the user to specify any time desired.

BUG: 477324
FIXED-IN: 6.1