kcm: Prepare DurationPromptDialog for submission to Kirigami
It can now handle more possible units. We keep the existing API,
but add unit options from milliseconds all the way to years in
addition to the existing two options for seconds and minutes.

Custom footer buttons are replaced with a standard OK/Cancel setup.
This renames the "Confirm" button to "OK", but that's not a bad
thing in terms of consistency with other dialogs.

The subtitle property is renamed to "label", which more accurately
describes its role and expresses the desire for a trailing colon.

Implicit width is specified such that a dialog without label will
still be large enough to fully display its header and footer,
with the main control horizontally centered in the dialog.

There is now only a single suffix label which picks its unit text
dynamically, and tries to keep its width static (via TextMetrics)
even in the fact of changing pluralization. It won't be perfect
for all languages, but should be an improvement at least.

The file is relicensed from GPL to LGPL. Both myself and the file's
original author, Kristen McWilliam, agreed to this change.