kcm: Factor out ComboBoxWithCustomValue for later reuse
Cuts down on a lot of boilerplate for the code that's using
the component. Likely also a better separation of concerns.

Model changes at this time are only monitored via onModelChanged,
so there is room for improvement still. For this use case though,
that limitation won't pose a problem.

Hoping to submit the component to Kirigami or another library of
common components, to make use of it in other KCMs as well.