feat: enable use as a screensaver
The screen locker is already in essence a screensaver, allowing one to
hide the unlock controls and clock, show various types of wallpaper
slideshows, etc. This is great since screensavers are almost
non-existent these days, and they are still wanted by many people.

The only thing missing really is the ability to use it as a screensaver
_without_ requiring a password to unlock.

FEATURE: 442092
FIXED-IN: 6.1