ksldapp: Set componentDisplayName property for global Lock Session action
This is used by kglobalacceld as the "friendly name" of a given component,
which kcm_keys then uses as a name in the Shortcuts KCM list.

When not set, the application name is used. Although different plugins and
libraries (e.g. plasma-workspace's startkde/session-shortcuts and
KScreenLockerSettings here) register global shortcuts for the "ksmserver"
component with the "Session Management" display name, KSldApp::initialize()
might be the first to register a ksmserver global shortcut.

When testing the latest Neon stable release (neon-user-20250119-0743) with a
new installation, kwin_wayland would often be the first to call
KSldApp::initialize(), so we would end up with a new component called
"ksmserver" with a friendly name of "KWin".

In practice, we would end up with two KWin components in the Shortcuts KCM,
the real one and the one that should be called "Session Management", which
has "Lock Session", "Log Out" etc.

CCBUG: 498683