X11Locker: always call stayOnTop when receiving XCB_MAP_NOTIFY to fix black screen
If a window was shown up previously and is hidden for now with the
background window together, m_allowedWindows will contain the window.
When the window shows up again, it will also raise the background
window. But when handling XCB_MAP_NOTIFY event from the background window,
stayOnTop is not called, so the background window becomes the top window.

CCBUG: 483163