kcm: make long wallpaper plugins scrollable
Wallpaper plugins can (and usually do) have their own settings, which
are dynamically loaded. There is a problem though because some wallpaper
plugins are based around a resizable view (such as a gridview of images)
that provides its own scollview and expect to fill the available area,
while others just provide items and rely on the loading container to
provide scrollability.

Instead, we set the implicitSize of the loader to its content and adjust
the main layout height accordingly if necessary (so (implicitly) large
items will cause scrollability, but keep the loader filling the layout
(so  resizable views take up the whole space). Along the way, we fix a
warning during the loading of the kcm.

CCBUG: 489580