kcm: set fewer unused config properties
When loading a wallpaper plugin configuration, all config keys of the
plugin are set initially, even though most plugins do not use all of
them, in particular not the properties that show the default values.
Unfortunately we cannot set them after loading, as some plugins may
require the values to be initially present, and we should not
unnecessarily break third-party plugins.

Instead we use the same strategy as in the wallpaper kcm and the desktop
containment configuration and load a temporary version of the plugin
configuration to determine which keys we need to set when loading it for
real. Along the way, we also set the wallpaperConfiguration property, as
the Image/SlideShow plugins require it, as we do in kcm_wallpaper.