greeter: Lower layershell layer to Top
We have a wayland protocol for requesting windows to be raised over the
lockscreen. However, the greeter's layershell layer is Overlay, which
makes it above other windows (though this doesn't seem to be consistent, seeing as the phone app works fine?). Lower the layershell layer of the greeter so that other
windows can be drawn over it.

Needed for https://invent.kde.org/plasma/plasma-mobile/-/merge_requests/270/