autotests: Make testRestartIdlePeriodAfterInhibit more robust
The intention of the test is to ensure we're locked 2 seconds after
uninhibiting. Due to the deliberate nature of coarse timer this can
drift quite a lot. Rather than doing multiple waits and checking the
state in between, this checks the time till the state changes with a
coarser check.