Drop defunct wallpaper from fallback theme
This is broken for the default setup because of a bug
in wallpaperPathForSize(), it doesn't return any wallpaper.

While that would be fixable in principle this is the only usage
of this API, and it's a nice-to-have optical thing for a
fallback that should never happen anyway.

As such, drop the wallpaper. To make the visuals slightly nicer
make the background Plasma-blue instead of black